# config/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: mobile-prepaid-card-service-provider
host: 'localhost:8000'
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /order:
    x-swagger-router-controller: Create new mobile prepaid order
    post:
      description: Create a mobile prepaid order
      operationId: createOrder
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/OrderParameters'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/OrderResponse'
        default:
          description: Error Occured
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/order/{id}':
    x-swagger-router-controller: Get order item by order id
    get:
      description: Check order info by order id
      operationId: getOrderItem
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/OrderId'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/OrderResponse'
        default:
          description: Error Occured
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/order/{id}/status':
    x-swagger-router-controller: Get order item status by order id
    get:
      description: Check order status by order id
      operationId: getOrderItemStatus
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/OrderId'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/OrderStatusResponse'
        default:
          description: Error Occured
          schema:
            $ref: '#/definitions/ErrorResponse'
  /provider/inquiry:
    x-swagger-router-controller: Count number of cards
    get:
      description: Count number of cards in store
      operationId: countCards
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/CountCardParameters'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/CountCardResponse'
        default:
          description: Error Occured
          schema:
            $ref: '#/definitions/ErrorResponse'
parameters:
  RequestId:
    name: request_id
    in: query
    type: string
    description: The request id to create an oder.
    required: true
  OrderId:
    name: id
    in: path
    type: string
    description: The order id to receive the order item.
    required: true
  OrderParameters:
    name: body
    in: body
    description: The parameters for creating new order
    schema:
      $ref: '#/definitions/OrdersRequest'
  CountCardParameters:
    name: query
    in: body
    description: The parameters for counting number of cards
    schema:
      $ref: '#/definitions/CountCardRequest'
definitions:
  OrdersRequest:
    required:
      - request_id
      - quantity
      - amount
      - opcode
    properties:
      request_id:
        type: string
        example: 434c9cb53bcf4e8e9b03ccdc18884c4d
      quantity:
        type: integer
        example: 1000
        description: The number of inquired cards
      amount:
        type: integer
        example: 100000
        description: The price unit per card
        enum:
          - 10000
          - 20000
          - 50000
          - 100000
          - 200000
          - 500000
      opcode:
        type: string
        example: VTT
        description: The network operation code such as VTT|VNP|VNN|VMS
  CountCardRequest:
    required:
      - request_id
    properties:
      request_id:
        type: string
        example: 434c9cb53bcf4e8e9b03ccdc18884c4d
      amount:
        type: integer
        example: 100000
        description: The price unit per card
        enum:
          - 10000
          - 20000
          - 50000
          - 100000
          - 200000
          - 500000
      opcode:
        type: string
        example: VTT
        description: The network operation code such as VTT|VNP|VNN|VMS
  OrderResponse:
    required:
      - request_id
      - order_id
      - status
      - items
    properties:
      request_id:
        type: string
        example: 434c9cb53bcf4e8e9b03ccdc18884c4d
        description: The request id
      order_id:
        type: string
        example: 434c9cb53bcf4e8e9b03ccdc18884c4d
        description: The reference id as known as order id
      status:
        type: string
        example: SUCCESS|PENDING|FAILED
        description: The code which reflect to request status
      items:
        type: array
        items:
          $ref: '#/definitions/OrderItemResponse'
      message:
        type: string
        example: Unknow Error
        description: The message reflect to status if status is FAILED
  OrderStatusResponse:
    required:
      - request_id
      - order_id
      - status
    properties:
      request_id:
        type: string
        example: 434c9cb53bcf4e8e9b03ccdc18884c4d
        description: The request id
      order_id:
        type: string
        example: 434c9cb53bcf4e8e9b03ccdc18884c4d
        description: The reference id as known as order id
      status:
        type: string
        example: SUCCESS|PENDING|FAILED
      message:
        type: string
        example: Unknow Error
        description: The message reflect to status if status is FAILED
  OrderItemResponse:
    properties:
      opcode:
        type: string
        example: VTT
      code:
        type: string
        example: 6af6b3fe090e4ae4
      amount:
        type: integer
        example: 100000
      serial:
        type: string
        example: 53835832354
      expire:
        type: string
        example: '2050-12-17 00:00:00'
  CountCardResponse:
    required:
      - request_id
      - order_id
      - amount
      - quantity
      - opcode
    properties:
      request_id:
        type: string
        example: 434c9cb53bcf4e8e9b03ccdc18884c4d
      order_id:
        type: string
        example: 434c9cb53bcf4e8e9b03ccdc18884c4d
        description: The reference id as known as order id
      quantity:
        type: integer
        example: 1000
        description: The number of inquired cards``
      amount:
        type: integer
        example: 100000
        description: The price unit per card
        enum:
          - 10000
          - 20000
          - 50000
          - 100000
          - 200000
          - 500000
      opcode:
        type: string
        example: VTT
        description: The network operation code such as VTT|VNP|VNN|VMS
      message:
        type: string
        example: Unknow Error
        description: The message reflect to status if status is FAILED
  ErrorResponse:
    required:
      - code
    properties:
      code:
        type: string
      message:
        type: string